taxii2/src/requirements.txt
antlr4-python3-runtime==4.8
certifi==2020.12.5
chardet==3.0.4
datefinder==0.7.1
idna==2.10
pika==1.1.0
pycti==4.0.1
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
simplejson==3.17.2
six==1.15.0
sseclient==0.0.27
stix2==2.1.0
stix2-patterns==1.3.2
taxii2-client==2.2.2
urllib3==1.26.2
